A more stable and compatible dynamic relationship addition
*** THIS IS STILL A HACK ***

I changed the relationship type from has_one to might_have, as the schema
layout suggests this to be the correct thing to do. Please change it back
if I am mistaken.

This should take care of both immediate problems - the inability to use the
relationship in question with prefetch, and the exceptions on some codepaths.

If either of the problem persists - tests should be written by Andrew Beverly
against the GADS repository itself, so that ribasushi can check what exactly
is going on