Fix invalid field ID errors when creating view of curval field
This was happening because only the permissions of the current layout
were being added to each layout, not all columns, even though all
columns are available in a single layout.

Tests needed to be updated to not use the superadmin for filter tests,
and also to use GADS::Instances

In addition, it needed to be ensured that the $filter hash was not
modified when constructing the search, as that meant some tests were not
idempotent.