Fix some timeline items not being retrieved
I think this is the right fix, but I am not sure why the addition and
subtraction of one day was in the code in the first place. From what I
can tell, the min and max range of the current items is retrieved, so
therefore any new items should be up to but not including those dates.
There should be no need for further adjustment as was the case, as this
was missing items with single day date ranges.