Fix curval fields with typeahead downloading all values
In edit.tt the code calls for all curval field values even if it is a
typeahead. This commit returns zero values server-side in such cases
rather than try and return them all. This is assuming that the values
function is not needed elsewhere - if it is then edit.tt will need to be
chagned to not call .values on a curval typeahead.