Ensure consistency when bulk downloading records
Before this commit, it was possible that large downloaded record sets
would be inconsistent. The reason is that data was fetched in chunks
using the database's paging, but if data changed inbetween fetches then
different groups of records would be retrieved. This commit takes a
snapshot of the record IDs first, and then downloads those in chunks.