Set explicit font-size on <body>  to prevent small fonts in IE
Respond.js does some magic to add a media query polyfill, and by doing so
it sets the font-size of <body> to 100%. This will result in way too small
fonts on the page. Setting an inline style to match our css (15px) will
instruct respond.js to return that value after it's magic is done.