Fix update of alert cache of view without curuser for all_users
If the alert cache was being updated for all users with a view that did
not have a curuser set, then an exception was occurring because @users
had only a null user in.

As well as that problem, the previous code had a rather strange
arrangement which looked like it was incorrectly running a deletion on
the alert cache for anything that wasn't the current user when updating
a view that was not a curuser view.

Bottom line, I'm unsure why the code was as it was, and this commit
definitely fixes at least one bug.