Table management testing
This will make it easier to add and remove records in due course and to
test filters against the test table to create views.

Note, currently the 'create_view.t' script needs the GADS_GROUPNAME
environment variable set to the human readable name of a group to assign
permissions to for the test table.  The user specified by GADS_USERNAME
should belong to this group.