Improve downloads of timeline
This attempts to make it easier for users to download and print the
timeline. It removes the need to specify size - the latest version of
timeline.js does a much better job of autosizing. It also provides a PDF
option. This is set to download at A3 landscape. Attempts to make this
also auto-size didn't work, so instead this should download as always A3
wide and then will produce as many pages as needed long.

This also required a customised version of Bootstrap CSS, to stop it
removing all colour from a page. Use http://getbootstrap.com/customize/
and deselect "print media styles". See
https://stackoverflow.com/questions/33410724/bootstrap-print-css-removes-background-color