Render timeline PDF better
The previously calculated figure for the width of the timeline PDF
didn't seem to work, at least with the latest stable version of
PhantomJS. The updated version seems to work, although purely from trial
and error - it doesn't seem to match what are defined values for the
default of 75dpi. Setting the DPI in PhantomJS is supported in the
latest master branch but not the latest binary.