Further attempts to improve timeline PDF render
The previous timeline rendering was still hit-and-miss for
correctly-sized PDFs. Various attempts were made to use DPI
specifications for PhantomJS, all without success.

This commit removes the attempt to specify the actual width of the
timeline as an absolute value, and instead specifies it as 100%, along
with reducing the zoom so that more fits on the page.