Track whether all fields have been built for a curcommon
This commit attempts to improve the situation around retrieving all
fields for a curcommon datum. Before this commit, all fields would be
retrieved whenever needed, without a note kept that it had been done.
This commit attempts to make the all fields retrieval a persistent flag.
This means that if/when they are all needed again, that they are not
unnecessarily all retrieved again.