Try and reduce the number of columns retrieved for a curcommon field
This is to deal with a problem, whereby all of an autocur field's
columns were retrieved, if that autocur field was part of a curval-edit
value. This meant that if a curval-edit value was edited within a large
record, and that curval-edit contained an autocur field, then all of
that large record's values were retrieved during the edit.