Fix reporting of messages during alert sending process
Before this fix, messages (including fatal ones) that were reported
during the alert sending process after a record were not written. This
is because the attempt to "unhide" the parent reporting try block was
unsuccessful. This is because when the parent try dispatcher was
retrieved, it evaluated to false (even though it existed), so its
reporting was never changed.