Allow curvals to refer to same table
This commit allows curvals to refer to the same table, thereby allowing
records to refer to other records in the same table.

This removes the previously-implemented recursive checks. I believe
these are no longer necessary as there are other checks in the code to
prevent recursive loops when building values, but monitoring will be
required.