Fix search on person field
Prior to this commit, searching on a person field using a view saved a
person's ID to the filter. This didn't work, as the filter expected a
textual value of the person. On the other hand, searching on a createdby
field expected the ID number. This commit makes the searches
consistently use a textual value.

This meant a change to the way that the internal createdby field was
implemented, as it is now necessary to join on that field's full
database table.