Fix curval items being added to timeline when curval is group
This fixes a problem whereby if a curval field is being used for
grouping on the timeline, then the curval's values still appeared on the
label for each item. Normally, the field that is used for grouping
should not appear on the label as it is already displayed as the group.