Reinstate contains filter functionality
This was originally removed, as it was thought to be causing very slow
queries, due to the inability to use an index on the search. However, it
now seems that the very slow queries were caused by poor join
optimsation, and that even indexable searches probably don't use an
index due to the complicated join criteria.