Deepen inclusion of all values and sub-values for code parameters
Up until now, only a limited depth of parameters were passed to a code's
parameters for evaluation. For example, if passing a curval value, it
would not include further curvals. This was to prevent infinite
recursion. This commit adds recursion checking, to include all values
up until ones that have already been included.