Fix missing timeline items in middle of range
This commit fixes a problem whereby some timeline items fell in a gap
between the start of default retrieval time (now) and the range going
back (starting a day earlier).

It also tidies up the calculation of retrieval ranges and the items
retrieved, which has resulted in these being slightly different values.