Cope with jQuery element clicking
I don't understand exactly why we need this, or how to work around it in
a better way, but the addition of jQuery to the record view page means
that a standard WebDriver click() on the "Delete" button no longer
works.

Calling jQuery's click() function by passing JavaScript code to
WebDriver does work, though, so that's what we do for now.

It's possible that using WebDriver's Actions API would do this in a
better way, without calling jQuery, but I couldn't get anything working
quickly:
://www.w3.org/TR/webdriver2/#actions