WebDriver tests for all records shown
Before creating a view, test all records are shown.  This means we need
to look for a different attribute (views use "aria-label", all records
use "data-thlabel") and filter out the default fields.

I could have filtered out those fields using Test::Deep, but in keeping
with the Test2 approach used in Test::GADSDriver, I decided not to.
However, I couldn't figure out how to check a subset of fields easily
using Test2 so I've commented on that as something to do in future.