Avoid running tests that can't pass
On loading, the GADS module reads its configuration and connects to the
configured database.  If no configuration file exists, this always
fails.

This makes it possible to run most of the test suite against a fresh
checkout and for it to pass in the absence of any problems.  Tests no
longer always fail in the absence of a config.yml and a deployed,
running database.