Add experimental Travis-CI configuration
This only runs the unit tests in the "t" directory for now.

Avoid Lua version 5.3 because it emits numbers as "10.0" when the test
suite expects them as "10".  Version 5.2 does what the test suite
expects.  I don't know Lua, but I suspect someone who does would
understand how to deal with this properly.