Fix views with curval pushed onto next page
This fixes a problem whereby with a curval in a certain position in a
view, the curval was pushed onto the next page of results in one part of
the records retrieval but not another (due to an unaccounted sort). This
resulted in a missing search param for record_later.