Fix user has access to incorrect fields when no view
This commit fixes a problem whereby when there is no view defined, a
user has access to all fields in the records view regardless.

The problem is inconsistent handling of permissions. When a record is
loaded for editing, all fields need to be populated so that they are
written on save. However, for any sort of view data operation, the
fields need to be restricted.

This aspect could be improved better. This commit will suffice for the
time being.