Fix unnecessary retrieval of filtered curval snapshot fields
This patch means that when a filtered curval is rendered in a record
form, that other unrelated filtered curval snapshot fields will not be
rendered. This was happening because all fields in a record were having
their values checked and evaluated, instead of just the ones related to
the field clicked on.