Work around a timing problem
The "delete a record" modal has the CSS class "fade" assigned to it,
which causes Bootstrap to use transitions to make the modal animate
gradually into view.

This causes WebDriver to try clicking on the element sometimes before it
is fully ready in the view.  This fails.  It might be possible to track
the element somehow to wait until it has fully animated, but disabling
the animation altogether during testing seems like an easier approach.