Fix timezone problems
This commit fixes a problem whereby the created time of a record was
being written to the database in local time. Subsequent versions were
being written in UTC. This led to a problem whereby a later version
could have an earlier timestamp, and therefore the wrong version was
being retrieved as the latest.

This commit ensures that times are always written to the database in
UTC, and are also processed in the application as UTC. They are only
converted to the local timezone for presentation and for calculated
fields. Additional timezones could be added in the future.