Fix incorrect alerts being sent with user view limits in place
This fixes a tricky bug whereby if the user's view limit was such that
the number of records they could see was equal to the number of records
in the alert being processed, then huge numbers of incorrect alerts
would be sent. This is because in the above condition, the current_id
search parameter would be missing, thereby assuming that alerts would be
sent on all records. This was particularly noticeable if the user only
had access to one record (probably the one they had just created) in the
case of processing an alert on a single record change.