Fix problem with fields having missing values if record has curval referring to same table
This commit fixes a problem for records that contain a curval referring
to the same table. The problem was that if the curval contained a
multi-value field, then that multi-value would first be fetched for the
curval's values. Then, when the same field should have been fetched for
the main record, it wasn't as it was flagged as already having been
fetched.

This was only a problem when the multivalue field was fetched after the
curval, hence the additional testing with reordered fields.