Fix autocurs being filtered with view extra limits
This is the same as a bug that was recently fixed for curvals, except
this commit is for autocurs. It meant that if the table that the autocur
referred to had a default view extra limit applied, that would also be
applied to the autocur when it shouldn't have been.