Fix unable to bulk update records unless user has all permissions
This commit fixes a bug when bulk updating records, whereby the error

Can't call method "dependent_not_shown" on an undefined value

would be thrown if a user didn't have read access to all fields in a
record, even if that field wasn't being updated by the user.

This commit changes the obligation to show only fields that can be
accessed by a user from the columns_selected() function to the
columns_render() function. The former returns all fields regardless of a
user's access, the latter only returns what a user has access to.