Fix another "invalid curval" error
This commit fixes a problem whereby a curval-edit field in a record
might contain another curval value refering to a table with limited
viewing rights. If those limited viewing rights were based on the value
of the main record, then when the main record was being written, it may
not have written the values that allow the user access to it.

Whilst this is potentially a bigger problem, the fix in this instance
was simply not to use values() to build the blank value. This was
expensive anyway.

In the longer-term, there may be the requirement to validate curval IDs
when writing. In this case, if they are limited based on the parent's
values, then further work would be required to ensure that only a
fully-written record was used, not the one in the process of being
written. Either that, or the dependencies need to be updated to write in
the correct order.