Fix retrieval of different types of multivalue calc values
This fixes a problem whereby if there were 2 multivalue calc fields in a
table with different return types, then one would be blank.

The change is to only retrieve a single field's multivalues at a time.
Some time ago all fields of the same type were retreived at the same
time, but this has been gradually eroded and is not worth it any more.