Fix a Webdriver test failure
Both the "Manage fields" and "Edit field" page have the "layout" class
set on the body element, so checking for this alone does not distinguish
between the two pages.  Consequently, sometimes the regular expression
match on the h2 element happened when still on the "Manage fields" page.

The "Edit field" page contains an element with the ID "basic-panel" and
the "Manage fields" page does not, so waiting for this element to appear
ensures "Edit field" is the current page.