Add chronological record view functionality
This commit adds a feature to show the chronological edits of a record.
It retrieves all versions of a record and then compares them one-by-one.

For curval values that can be edited within the record, it relies on the
version times of the records matching. As such, to work on existing
edits, the times will need to be updated to match (until this commit the
times would differ slightly depending on the exact time each record was
written)