Fix memory leak
This commit fixes a problem with circular references during GADS::Layout
cloning. The clone (created within a field) contained a reference back
the layout which then contained a reference to the same field. This
commit ensures that any access to a GADS::Column is always done back to
the original "parent" layout object