Fix unnecessarily-large database queries during alert processing
This commit fixes a problem whereby large SQL queries were being
generated during the searching of views (much larger than the view
itself). The problem was that GADS::Records was being reused, which
meant that the joins of the previous views were being used in each
iteration of view search.