Fix error when table has 2 autocurs
This fixes a problem when a value is added to a curval field, when that
curval field has references back to 2 parent tables (the one that the
curval has been added to, and another one). The previous code assumed
that the autocur field would always refer back to the same table, and
prepopulated the value as such. This ensures that the value is only
prepopulated if it matches the same table it came from.