Attempt to fix problems with display conditions on snapshot fields
A problem has been identified whereby fields that have a display
condition dependent on filtered snapshots (filval) are not showing
correctly, as the filval field is empty on page render. Ordinarily the
flag in the field from the server should dictate whether they are being
displayed, but this is being overridden by a subsequent change triggered
when the widget is rendered on page load.