Fix exception when a table has 2 autocur fields to same table
This commit fixes a problem when a table has 2 autocur fields referring
to the same table. Before this patch, the already_seen variable would
remove the subfields of the second autocur field, thereby not retrieving
its values. This happened during the evaluation of associated calc
fields that used the autocur values.

This patch introduces the use of a Tree::DAG_Node to track the full
addition of curval/autocur fields, which whilst slight overkill does
make things easier and may be used in the future for additional similar
tracking.