Performance - further reductions on record evaluation where possible
This commit makes a very small change, by looking in the calc fields of
dependent autocurs for the name of changed fields, rather than just
including all fields with a short name regardless.

The actual line change is very small, just this:

    next unless grep $code =~ /\Q$_\E(?![_a-z0-9])/i, map $_->name_short,
        grep $_->name_short, @columns_changed;

But the overall commit includes some tidying up and the generation of
@columns_changed. This is used last to ensure that all changed fields
are included.