Fix validation of required file fields
Copying from associated commit in roger styleguide:

This commit fixes problems with validating file input fields that are
required. The problem was caused by files being immediately moved into
checkboxes when uploading, such that the HTML file input never actually
contained a file. This commit changes the required checks to be on the
checkboxes rather than the HTML file input.

It also requires some associated changes to the HTML templates, which
will be undertaken separately in the GADS source tree.