Fix HTML entities in records table headers
This major commit primarily fixes problems with including HTML entities
in the table header of the records table.

The rewrite required passing the datatables config to the JS code as an
encoded base64 string rather than raw JSON string. The latter was
causing problems with encoding and passing HTML entities as paramters.