Added keyboard funtionality to pop-over
Noticed that the curval pop-overs were not selectable with keyboards, so added the option trigger event with space or enter in addition to the standard click.

Also added check to make sure current target isn't a button. As selecting buttons within the row (via keyboard) were also triggering this event by mistake.