Calculator button
Added calculator button, removed extra button component as no longer needed, and updated site.js

It is worth noting for now that in the original button component had a "hanging" function (i.e. not implemented) called `this.initCalculator()` - it is thought that this is why the calculator was inactive in v2. Discussion is required to ascertain whether implementation is required, or the button can be removed.