Added auto height functionality for textareas
Enables auto-resize for text areas with class 'auto-adjust'
adjustTextareaHeight() calculates the total height of the text-area content (by putting it in a hidden div.)

Then calculates the height of one line (in REM).

Then if the content is more than one line, it adds the calculated line height to the total height. If not then just default to total height.