Enable different components on same object
This fixes what appears to be a bug, whereby different components
are not able to be initialised on the same element. This is because once
any component is initialised on an element, a single data attribute is
set, which then causes any subsuquent components to be skipped on the
same object.