Fix duplicate trigger of change event for single-select
This commit fixes a problem whereby the change event was called twice
when a single-select option was changed.

This is because the change event was triggered from the updateState
function, and was also triggered automatically from the option itself
being changed. This was different to multiple select, which specifically
had stopPropagation called.

I cannot find any reason why stopPropagation is needed for the multiple
select, so in order to simplify things I have removed both the
stopPropagation from the multiple select and also the manual trigger
from updateState. As far as I can tell, this results in reliable change
events as would be expected (once for each time the widget value is
changed).