Fix ownerDocument is null error in table view
This commit fixes an error for a table containing blank elements within
more-less containers. The error was occurring because the blank elements
had zero height and were thus considered hidden, which meant that the
code traversed to the top of the DOM, eventually hitting the root
element which errored when css() was called on it.

This adds a few fixes, firstly bailing out early if the element is
blank, and also handling the above error gracefully should it happen .