Added all code referencing `self=this`
I have found that the datatables code threw a fit on use with the current changes, I've done some digging and found that the reference to this as self within the code is the issue due to me being incorrect about scoping in some places. I have fixed any scoping issues and self/this references within the code throughout.