Fix for filter typeahead querybuilder bug
A bug with unknown cause is meaning that dynamic loading of Typeahead as a chunk is causing filter typeahead on curvals to be rendered without rules across the board. For now, dynamic loading in this instance has been removed as a short-term solution, a longer-term solution will be pursued going forward.