New version
Took some doing to get files to behave as expected - a number of null references among other things seemed to be sticking around. Couldn't get the datum to work with the `_is_purged` sub/property, so I just made it work where I did for now.
If it is needed, I will work on it (`is_purged`) for all datum as a separate feature, however I believe for now this fits our requirements.