Remove a hardcoded username and password
The bin/setup_database has GADS_USERNAME and GADS_PASSWORD environment
variables to customise the username and password created.  The hardcoded
username and password added in 55e865f22c1 that creates a username and
password in every situation regardless of whether it's necessary.

Instead of using those hardcoded credentials, use the environment
variables to create a user for the Cypress tests.