Fix panics using deleted people in calcs
This fix parses the requisite datetime as a value from the database
rather than a (changing) CLDR format.

It has also required removing the standard :all import of
MooX::Types::MooseLike::Base as this was importing "Enum" which was
conflicting with the use of the GADS::Column::Enum class.

It also required the consistent use of GADS::DateTime has a role.